requirements/default.txt
django-ordered-model>=3.0
